inheritance fixed and classes being changed
i refactored my inheritance and created and abstract journey class that is the parent to emissionsdata and journeyrequest. I passed the calculated address to the createRequestBody method from the destinationId but for some reason I still cant get it to work. hmm
